% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osint_colours.R
\name{get_osr_hex}
\alias{get_osr_hex}
\title{Extract OSINTR colours as hex codes}
\usage{
get_osr_hex(...)
}
\arguments{
\item{...}{The given name/s of a OSINTR colour (eg 'astronaut blue' or 'iron')}
}
\description{
This function allows you to get the hex code associated with one or more OSINTR
colours. It's often used inside \code{osr_plot} to set specific colours for
certain series.
}
\examples{
get_osr_hex("persian green")
get_osr_hex("midnight", "mango tango")

}
