% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osint_colours.R
\name{make_osr_pal}
\alias{make_osr_pal}
\title{Interpolate a OSINTR colour palette}
\usage{
make_osr_pal(palette = "graph", reverse = FALSE, ...)
}
\arguments{
\item{palette}{(character; default = \code{"brand"}) given name of a OSINTR
palette: \code{\link{osr_palettes}}}

\item{reverse}{(boolean; default = \code{FALSE}) indicating if palette should
be reverse}

\item{...}{Additional arguments to pass to \code{colorRampPalette} see
details here \code{\link[grDevices]{colorRamp}}}
}
\description{
This function takes a OSINTR colour palette and generates more colours from it,
so that there are enough to make your chart.
}
\details{
The interpolation method is set to "spline" (the default is "linear") in an
attempt to reduce the number of vomit colours that get produced when
generating many colours.
}
\seealso{
\code{\link{osr_palettes}}
}
