% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osint_colours.R
\docType{data}
\name{null_plot_opts}
\alias{null_plot_opts}
\title{The null plotting options}
\format{
An object of class \code{list} of length 4.
}
\usage{
null_plot_opts
}
\description{
These options restore the default ggplot2 colours options for the R session.
}
\keyword{datasets}
