% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osint_package.R
\docType{package}
\name{osintr-package}
\alias{osintr}
\alias{osintr-package}
\title{osintr: Useful OSINT functions}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
