% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osint_colours.R
\docType{data}
\name{osr_colours}
\alias{osr_colours}
\title{The OSINTR colours}
\format{
An object of class \code{character} of length 22.
}
\usage{
osr_colours
}
\description{
These colours are used in \code{osr_plot} and form the basis for our
palettes.
}
\keyword{datasets}
