% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osint_colours.R
\name{osr_gmap}
\alias{osr_gmap}
\title{OSINTR-styled ggmap wrapper}
\usage{
osr_gmap(..., crs = 7844)
}
\arguments{
\item{...}{Standard ggmap arguments}

\item{crs}{Numeric CRS for \code{ggplot2::coord_sf} (default = 7844, GDA2020
degrees). See \url{https://epsg.io/7844} for others (7845 is for meters)}
}
\description{
If you get an error about not enough colours, try a different palette (such
as graph) or set \code{discrete = FALSE} in \code{scale_colour_osr} or
\code{scale_fill_osr}
}
