% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osint_colours.R
\name{osr_map_save}
\alias{osr_map_save}
\title{Save a map made with osr_gmap}
\usage{
osr_map_save(
  filename,
  plot = ggplot2::last_plot(),
  dpi = 320,
  scale = 1.5,
  height = 11.25,
  width = 18,
  device = "png",
  type = "cairo",
  units = "cm",
  ...
)
}
\arguments{
\item{filename}{(character) file name to create on disk.}

\item{plot}{The plot to save}

\item{dpi}{(numeric; default = 320) dpi to use}

\item{scale}{(numeric; default = 1.5)     multiplicative scaling factor}

\item{height}{(numeric; default = 11.25) image height}

\item{width}{(numeric; default = 18) image width}

\item{device}{(character; default = 'png') graphics device}

\item{type}{(character; default = 'cairo') type}

\item{units}{(character; default = 'cm') units}

\item{...}{Standard ggsave arguments}
}
\description{
Very similar to \code{osr_save}, but this will extract the aspect ratio of
the map passed to it and edit the image dimensions to accommodate that ratio.
}
