% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osint_colours.R
\docType{data}
\name{osr_palettes}
\alias{osr_palettes}
\title{Predefined OSINTR colours combined into palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
osr_palettes
}
\description{
This is a list of OSINTR colours combined into palettes. The palettes are used
for different plots and maps.
}
\keyword{datasets}
