% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osint_colours.R
\name{osr_plot}
\alias{osr_plot}
\title{OSINTR-styled ggplot2 wrapper}
\usage{
osr_plot(...)
}
\arguments{
\item{...}{Standard ggplot arguments}
}
\description{
If you get an error about not enough colours, try a different palette (such
as graph) or set \code{discrete = FALSE} in \code{scale_colour_osr} or
\code{scale_fill_osr}
}
\seealso{
\code{\link{scale_colour_osr}}

\code{\link{scale_fill_osr}}
}
