% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osint_colours.R
\docType{data}
\name{osr_plot_opts}
\alias{osr_plot_opts}
\title{The OSINTR plotting options}
\format{
An object of class \code{list} of length 4.
}
\usage{
osr_plot_opts
}
\description{
These options set the default ggplot2 colours options for the R session.
}
\keyword{datasets}
