% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osint_colours.R
\name{osr_save}
\alias{osr_save}
\title{Save images with opinions}
\usage{
osr_save(
  filename,
  plot = ggplot2::last_plot(),
  dpi = 320,
  scale = 1.5,
  height = 11.25,
  width = 18,
  device = "png",
  type = "cairo",
  units = "cm",
  bg = "white",
  ...
)
}
\arguments{
\item{filename}{(character) file name to create on disk.}

\item{plot}{The plot to save}

\item{dpi}{(numeric; default = 320) dpi to use}

\item{scale}{(numeric; default = 1.5)     multiplicative scaling factor}

\item{height}{(numeric; default = 11.25) image height}

\item{width}{(numeric; default = 18) image width}

\item{device}{(character; default = 'png') graphics device}

\item{type}{(character; default = 'cairo') type}

\item{units}{(character; default = 'cm') units}

\item{bg}{(character; default = "white") plot background colour, needed
because of update to \code{ggsave}}

\item{...}{Standard \code{ggsave} arguments}
}
\description{
Wrapper for \code{ggsave} using golden ratio, png and cairo and other
opinions. Using type = 'cairo' fixes some anti-aliasing problems.
}
\details{
All arguments passed to \code{ggsave()}
}
