% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osint_colours.R
\name{scale_colour_osr}
\alias{scale_colour_osr}
\alias{scale_fill_osr}
\title{Scale constructor for OSINTR colours}
\usage{
scale_colour_osr(palette = NULL, discrete = FALSE, reverse = FALSE, ...)

scale_fill_osr(palette = NULL, discrete = FALSE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in \code{osr_palettes}}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or
not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to \code{\link{discrete_scale}} or
\code{\link{scale_color_gradientn}}, used respectively when discrete is
TRUE or FALSE}
}
\description{
If you get an error about not enough colours, try a different palette (such
as \code{graph}) or set \code{discrete = FALSE}.
}
\section{Functions}{
\itemize{
\item \code{scale_colour_osr}: For colour scales

\item \code{scale_fill_osr}: For fill scales
}}

