% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osint_colours.R
\name{set_geom_defaults}
\alias{set_geom_defaults}
\title{Set default geom aesthetics}
\usage{
set_geom_defaults(geom, new)
}
\arguments{
\item{geom}{(character) the geom to update}

\item{new}{(various) the aesthetics to update}
}
\description{
All arguments are passed to \code{ggplot2::update_geom_defaults}, but this
function adds a \code{NULL} check.
}
