% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osint_colours.R
\name{set_osr_defaults}
\alias{set_osr_defaults}
\title{Set the default ggplot2 aesthetics to OSINTR's}
\usage{
set_osr_defaults()
}
\description{
A wrapper for a bunch of other functions to set ggplot2 default aesthetics.
Used on load.
}
