% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osint_colours.R
\name{theme_osr}
\alias{theme_osr}
\title{OSINTR-styled ggplot2 theme}
\usage{
theme_osr(...)
}
\arguments{
\item{...}{arguments pass to \code{theme_minimal}}
}
\description{
A ggplot2 theme based on \code{theme_minimal} using Calibri, OSINTR colours and
customised element positioning.
}
\details{
ggplot2 themes don't affect \code{geom_} colours, they are addressed
elsewhere.
}
