% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osint_colours.R
\name{theme_osr_map}
\alias{theme_osr_map}
\title{OSINTR-styled ggplot2 theme for maps}
\usage{
theme_osr_map(...)
}
\arguments{
\item{...}{arguments pass to \code{theme_bw}}
}
\description{
A ggplot2 theme based on \code{theme_bw} using Calibri, OSINTR colours and
customised element positioning.
}
\details{
ggplot2 themes don't affect \code{geom_} colours, they are addressed
elsewhere.
}
